Move attribute related fixtures from global fixture file
