+++ saleor/attribute/tests/fixtures/__init__.py
@@ -1,0 +1,2 @@
+from .attribute import *  # noqa: F403
+from .translation import *  # noqa: F403
